refactor: 사용자 입력한 값 유효성 체크하는 클래스 구현
